+++ R/read_write_pomdp_solve.R
@@ -88,7 +88,53 @@
   
   stop("Illegal belief format.\n", belief)
 }
+
+### FIXME... test translating a data.frame with T_
+.translate_transitions <- function(model) {
+  states <- model$model$states
+  actions <- model$model$actions
+  observations <- model$model$observations
+  trans <-  model$model$transition_prob
   
+  if(is.data.frame(trans)) {
+    tr <- lapply(actions, FUN = function(x) 
+      matrix(0, nrow = length(states), ncol = length(states), 
+        dimnames = list(states, states)))
+    names(tr) <- actions
+    
+    for(i in 1:nrow(trans)){
+      if(trans[i,"action"] == "*") acts <- actions
+      else acts <- trans[i,"action"]
+      for(a in acts) {
+        if(trans[i,"start.state"] == "*" && trans[i,"start.state"] == "*")
+          tr[[a]][] <- trans[i,"probability"]
+        else if (trans[i,"start.state"] == "*")
+          tr[[a]][, trans[i,"end.state"]] <- trans[i,"probability"]
+        else if (trans[i,"end.state"] == "*")
+          tr[[a]][trans[i,"start.state"], ] <- trans[i,"probability"]
+        else 
+          tr[[a]][trans[i,"start.state"], trans[i,"end.state"]] <- trans[i,"probability"]
+      }
+    }
+  } else if(is.list(trans)) {
+    trans <- lapply(trans, FUN = function(tr) {
+      if(is.character(tr)) {
+        tr <- switch(tr, 
+          identity = diag(1, nrow = length(states), ncol = length(states)),
+          uniform = matrix(1/length(states), 
+            nrow = length(states), ncol = length(states)))
+        dimnames(tr) <- list(states, states)
+      }
+      if(!is.matrix(tr)) 
+        stop("Transition probabilities cannot be converted to matrix.")
+      
+      tr
+    })
+  } else stop("Unknown transition matrix format.")
+  
+  trans  
+}
+
 ## helpers to read/write pomdp-solve files
 
 # alpha file is a matrix with # of states columns