+++ NEWS.md
@@ -11,6 +11,7 @@
 * added plot_value_function.
 * added function sample_belief_space to sample from the belief space.
 * added function plot_belief_space.
+* added function transition_matrix.
 
 
 # pomdp 0.9.2 (12/06/2019)